# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "MarketPlace Standard App API",
    "description": "Standard app API for the MarketPlace applications.",
    "contact": {
      "name": "The Materials MarketPlace Consortium",
      "url": "https://www.materials-marketplace.eu/",
      "email": "dirk.helm@iwm.fraunhofer.de"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "FrontPage"
        ],
        "summary": "Frontpage",
        "description": "Open the frontpage of the app.",
        "operationId": "frontend",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not found."
          },
          "401": {
            "description": "Not authenticated."
          },
          "500": {
            "description": "Internal server error."
          },
          "503": {
            "description": "Service unavailable."
          }
        },
        "security": [
          {
            "AuthTokenBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "AuthTokenBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}